+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Patch circular import in mdast-util-to-hast dependency (#99)
+-   Display intermediary state while checking credentials (#102)
 
 ### Changed